update compute_hooke_force with recent devs
