+++ src/polyhedra/compute_hooke_classify_interactions.cpp
@@ -40,225 +40,6 @@ under the License.
 #include <exaDEM/drivers.h>
 #include <exaDEM/compute_hooke_interaction.h>
 
-
-
-namespace exaDEM
-{
-	using namespace exanb;
-
-	struct hooke_law_polyhedra
-	{
-		template<typename Cells>
-			ONIKA_HOST_DEVICE_FUNC void operator()(Interaction& item, Cells& cells, const HookeParams& hkp, const shapes& shps, const double time, mutexes& locker)
-			{
-				auto get_r = [&cells] (const int cell_id, const int p_id) -> const Vec3d 
-				{
-					const Vec3d res = {
-						cells[cell_id][field::rx][p_id],
-						cells[cell_id][field::ry][p_id],
-						cells[cell_id][field::rz][p_id]};
-					return res;
-				};
-
-				auto get_v = [&cells] (const int cell_id, const int p_id) -> const Vec3d 
-				{
-					const Vec3d res = {
-						cells[cell_id][field::vx][p_id],
-						cells[cell_id][field::vy][p_id],
-						cells[cell_id][field::vz][p_id]};
-					return res;
-				};
-
-				// === positions
-				const Vec3d ri = get_r(item.cell_i, item.p_i);
-				const Vec3d rj = get_r(item.cell_j, item.p_j);
-
-				// === cell
-				auto& cell_i =  cells[item.cell_i];
-				auto& cell_j =  cells[item.cell_j];
-
-				// === vrot
-				const Vec3d& vrot_i = cell_i[field::vrot][item.p_i];
-				const Vec3d& vrot_j = cell_j[field::vrot][item.p_j];
-
-				// === type
-				const auto& type_i = cell_i[field::type][item.p_i];
-				const auto& type_j = cell_j[field::type][item.p_j];
-
-				// === vertex array
-				const auto& vertices_i =  cell_i[field::vertices][item.p_i];
-				const auto& vertices_j =  cell_j[field::vertices][item.p_j];
-
-				// === shapes
-				const shape* shp_i = shps[type_i];
-				const shape* shp_j = shps[type_j];
-
-				auto [contact, dn, n, contact_position] = detect[item.type](vertices_i, item.sub_i, shp_i, vertices_j, item.sub_j, shp_j);
-				if(contact)
-				{
-					const Vec3d vi = get_v(item.cell_i, item.p_i);
-					const Vec3d vj = get_v(item.cell_j, item.p_j);
-					const auto& m_i = cell_i[field::mass][item.p_i];
-					const auto& m_j = cell_j[field::mass][item.p_j];
-
-					// temporary vec3d to store forces.
-					Vec3d f = {0,0,0};
-					const double meff = compute_effective_mass(m_i, m_j);
-
-					hooke_force_core(dn, n, time, hkp.m_kn, hkp.m_kt, hkp.m_kr,
-							hkp.m_mu, hkp.m_damp_rate, meff,
-							item.friction, contact_position,
-							ri, vi, f, item.moment, vrot_i,  // particle 1
-							rj, vj, vrot_j // particle nbh
-							);
-
-
-					// === update particle informations
-					// ==== Particle i
-					locker.lock(item.cell_i, item.p_i);
-
-					auto& mom_i = cell_i[field::mom][item.p_i];
-					mom_i += compute_moments(contact_position, ri, f, item.moment);
-					cell_i[field::fx][item.p_i] += f.x;
-					cell_i[field::fy][item.p_i] += f.y;
-					cell_i[field::fz][item.p_i] += f.z;
-
-					locker.unlock(item.cell_i, item.p_i);
-
-					// ==== Particle j
-					locker.lock(item.cell_j, item.p_j);
-
-					auto& mom_j = cell_j[field::mom][item.p_j];
-					mom_j += compute_moments(contact_position, rj, -f, -item.moment);
-					cell_j[field::fx][item.p_j] -= f.x;
-					cell_j[field::fy][item.p_j] -= f.y;
-					cell_j[field::fz][item.p_j] -= f.z;
-
-					locker.unlock(item.cell_j, item.p_j);
-				}
-				else
-				{
-					item.reset();
-				}
-			}
-		typedef decltype (&(exaDEM::detection_vertex_vertex_precompute)) Detector;
-		const std::vector<Detector> detect = std::vector{ exaDEM::detection_vertex_vertex_precompute, 
-			 exaDEM::detection_vertex_edge_precompute,
-			 exaDEM::detection_vertex_face_precompute,
-			 exaDEM::detection_edge_edge_precompute};
-};
-
-// C for cell and D for driver
-template<typename TMPLD>
-struct hooke_law_driver
-{
-	template<typename Cells>
-		ONIKA_HOST_DEVICE_FUNC void operator()(Interaction& item, Cells& cells, Drivers& drvs, const HookeParams& hkp, const shapes& shps, const double time, mutexes& locker)
-		{
-			const int driver_idx = item.id_j; //
-			auto& driver = std::get<TMPLD>(drvs.data(driver_idx)) ;
-			auto& cell = cells[item.cell_i];
-			const auto type = cell[field::type][item.p_i];
-			auto* shp = shps[type];
-
-			const size_t p   = item.p_i;
-			const size_t sub = item.sub_i;
-			// === positions
-			const Vec3d r       = { cell[field::rx][p], cell[field::ry][p], cell[field::rz][p] };
-			// === vrot
-			const Vec3d& vrot  = cell[field::vrot][p];
-			// === vertex array
-			const auto& vertices =  cell[field::vertices][p];
-
-			auto [contact, dn, n, contact_position] = exaDEM::detector_vertex_driver(driver, vertices, sub, shp);
-
-			if(contact)
-			{
-				constexpr Vec3d null = {0,0,0};
-				auto& mom = cell[field::mom][p];
-				const Vec3d v = { cell[field::vx][p], cell[field::vy][p], cell[field::vz][p] };
-				const double meff = cell[field::mass][p];
-				Vec3d f = null;
-				hooke_force_core(dn, n, time, hkp.m_kn, hkp.m_kt, hkp.m_kr,
-						hkp.m_mu, hkp.m_damp_rate, meff,
-						item.friction, contact_position,
-						r, v, f, item.moment, vrot,  // particle i
-						driver.center, driver.get_vel(), driver.vrot // particle j
-						);
-
-				// === update informations
-				locker.lock(item.cell_i, p);
-				mom += compute_moments(contact_position, r, f, item.moment);
-				cell[field::fx][p] += f.x;
-				cell[field::fy][p] += f.y;
-				cell[field::fz][p] += f.z;
-				locker.unlock(item.cell_i, p);
-			}
-			else
-			{
-				item.reset();
-			}
-		}
-};
-
-struct hooke_law_stl
-{
-	template<typename Cells>
-		ONIKA_HOST_DEVICE_FUNC void operator()( Interaction& item, Cells& cells, Drivers& drvs, const HookeParams& hkp, const shapes shps, const double time, mutexes& locker)
-		{
-			const int driver_idx = item.id_j; //
-			auto& driver = std::get<Stl_mesh>(drvs.data(driver_idx)) ;
-			auto& cell = cells[item.cell_i];
-			const auto type = cell[field::type][item.p_i];
-			auto* shp_i = shps[type];
-
-			const size_t p_i   = item.p_i;
-			const size_t sub_i = item.sub_i;
-			const size_t sub_j = item.sub_j;
-
-
-			// === positions
-			const Vec3d r_i       = { cell[field::rx][p_i], cell[field::ry][p_i], cell[field::rz][p_i] };
-			// === vrot
-			const Vec3d& vrot_i  = cell[field::vrot][p_i];
-			const Quaternion& orient_i  = cell[field::orient][p_i];
-			const auto& shp_j = driver.shp;
-
-			const Quaternion orient_j = {1.0,0.0,0.0,0.0};
-			auto [contact, dn, n, contact_position] = func(item.type, r_i, sub_i, shp_i, orient_i, driver.center, sub_j, &shp_j, orient_j);
-
-			if(contact)
-			{
-				constexpr Vec3d null = {0,0,0};
-				auto& mom = cell[field::mom][p_i];
-				const Vec3d v_i = { cell[field::vx][p_i], cell[field::vy][p_i], cell[field::vz][p_i] };
-				const double meff = cell[field::mass][p_i];
-				Vec3d f = null;
-				hooke_force_core(dn, n, time, hkp.m_kn, hkp.m_kt, hkp.m_kr,
-						hkp.m_mu, hkp.m_damp_rate, meff,
-						item.friction, contact_position,
-						r_i, v_i, f, item.moment, vrot_i,  // particle i
-						driver.center, driver.vel, driver.vrot // particle j
-						);
-
-				// === update informations
-				locker.lock(item.cell_i, p_i);
-				mom += compute_moments(contact_position, r_i, f, item.moment);
-				cell[field::fx][p_i] += f.x;
-				cell[field::fy][p_i] += f.y;
-				cell[field::fz][p_i] += f.z;
-				locker.unlock(item.cell_i, p_i);
-			}
-			else
-			{
-				item.reset();
-			}
-		}
-
-	const stl_mesh_dectector func;
-};
-}
-
 namespace exaDEM
 {
 	using namespace exanb;
@@ -287,8 +68,6 @@ namespace exaDEM
 			return R"EOF(
                 )EOF";
 		}
-
-
 
 		template<typename... Args>
 			inline void run_hooke_law(int id, Classifier& ic, Args&&... args)
